Remove MakeRFC2822DateString().
This function is called in one place. By implementing directly there, we
can use a simpler StringPrintf()-based implementation, and in the future
an even simpler implementation from time_formatting.h.

This also simplifies the test to not hardcode the expectation that the
timezone offset is "-0000". The RFC doesn't require this, it's more
correct to report the actual local offset, and in the future we will do
so. For now the behavior is unchanged, but the test doesn't need this to
pass.

Bug: 1476086
Change-Id: I0d6000b78df94e477a077e408a7269547b25b0b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4867446
Commit-Queue: Justin Novosad <junov@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197197}
NOKEYCHECK=True
GitOrigin-RevId: f6bb548d4853f2adc3214eecdd63561fe912799f