Change old fast counters tests
Test changes are due to the:
https://drafts.csswg.org/css-lists/#inheriting-counters
which says that we inherit the counter from sibling, only
if element counters does not already contain a counter
with the same name. But as per:
https://drafts.csswg.org/css-lists/#counter-inheritance-example
the value of the counter should be taken from previous element
in tree order. That's why the set for counters contain only 1 counter.

Bug: 990657
Change-Id: I7bc6b6b05ea61b3d13ff27b2311d8234bd4624ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4834477
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197773}
NOKEYCHECK=True
GitOrigin-RevId: c63b6a459170c6e72a93bb7c66577ad23a1a060b