Merge and rename OnFrameDropped+OnFrameDroppedInternal.
As of [1], we no longer need two methods and can merge them into a
single method, because all OnFrameDropped was doing was calling
OnFrameDroppedInternal.

While we're at it, we rename it to OnFrameDroppedInRenderer to make the
name of the event less ambiguous.

This is in anticipation to a future CL that will forward frame drops
events from the browser process to the renderer process, so we'll soon
have an OnFrameDroppedOutOfProcess method that tells the track that
a frame was dropped earlier in the pipeline as opposed to the other way
around (OnFrameDroppedInRenderer is the track forwarding an event back
to the browser process for UMA reasons).

[1] https://chromium-review.googlesource.com/c/chromium/src/+/4843050

Bug: chromium:1481448, chromium:1472978
Change-Id: Id0dfbbcaec661927c7a323c4816886fe6821f3c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4866320
Reviewed-by: Tony Herre <toprice@chromium.org>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197865}
NOKEYCHECK=True
GitOrigin-RevId: 8e67454a64f8800c96ef0b99a50636615a78f9ef