[Exo] Add a new request to apply rounded corners to the window for remote shell
A new protocol was added for aura_shell that allows the client to
request a rounded window in this CL: crrev.com/c/4775627

One of the client to use this protocol is ARC++. ARC++ is migrating
from remote-shell to aura-shell, but the work is in progress so
we still need to use remote-shell for the old ARC++ versions.

Therefore to support rounded windows for older ARC++ versions, porting
the protocol to remote shell as well.

Bug: 1415486, b:287060338
Change-Id: I6a953f33dbe09d248505be432fd3f33b31b7e4f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4869221
Commit-Queue: Zoraiz Naeem <zoraiznaeem@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197927}
NOKEYCHECK=True
GitOrigin-RevId: 65d66418536059713641efecc3b02bde1e027241