Make AXObject::IsKeyboardFocusable use Element::IsKeyboardFocusable
This was previously a re-implementation of Element::IsKeyboardFocusable,
but after recent refactoring, we should just be able to directly use
the Element version.

In implementing this, I found another semi-issue with
IsKeyboardFocusable on a number of element types: it did not take
`tabindex` into account. That wasn't a problem in practice while
navigating with the keyboard, since focus_controller (at [1]) also
looked at the `tabindex` value and took it into account. But that did
mean `IsKeyboardFocusable()` didn't match the actual focusability.
With this change, it does.

This also cleans up the comment around Element::SupportsFocus().

Fixed: 1444450
Change-Id: I8a372fcb9542ecc56c90c15ebb8dc69fbb0c9075
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4844436
Commit-Queue: Mason Freed <masonf@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198050}
NOKEYCHECK=True
GitOrigin-RevId: f33685b561f5319561ed75d572a58b5763a718d8