[cc] Simplify and adjust memory limits on Android
On Android, compositor memory limits are derived from system memory and
dalvik memory limits. This code was noted as outdated in 2017 (see
linked bug), and as a result didn't work the way it should have.

This is because:
- Reported system RAM on Android is lower than installed RAM because of
  carveouts
- Dalvik memory limits are not really correlated with system RAM

Based on field data, the large majority of devices is running with a
computed limit of 256MiB, with some devices using 96MiB. This CL
simplifies the code to:
- Low-end or <2GiB: 96MiB
- Otherwise, 256MiB

Which mostly matches the current in-the-wild reality. These limits are
likely not optimal, but at least this simplifies the code. Also, lower
the priority cutoff to "NICE_TO_HAVE", which matches desktop, and
reality, since ALLOW_EVERYTHING is lowered to NICE_TO_HAVE elsewhere in
the code (see PriorityCutoffToTileMemoryLimitPolicy() for instance).

This is gated behind a feature flag, to make sure that this is not
breaking things.

Bug: 742534
Change-Id: Ie9525f4a18a6a7057a5e39734f50db3a7ccd0fd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4860245
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198269}
NOKEYCHECK=True
GitOrigin-RevId: 42b97eaa9deace1c188434ccac37ee22223ade4f