[v8 compile hints] When streaming, check data production stricter
This CL is pre-work for the CL which makes the "should we produce
data" decision based on client-side randomization, not based on a
Finch-controlled flag.

In that case, we need V8CrowdsourcedCompileHintsProducer to do the
randomization, so the streaming code path needs to check
V8CrowdsourcedCompileHintsProducer too, checking the flag won't be
enough.

Bug: chromium:1406506
Change-Id: I53159d80fb547f95a3cd58ce190040b7b4b02e5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4872716
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198276}
NOKEYCHECK=True
GitOrigin-RevId: a0b7dad3b5fb606bcf676418d1d91c45cd9da98f