[DL] Ensure grid placement data is dirty.
After:
https://chromium-review.googlesource.com/c/chromium/src/+/4859555

We started to return nullptr from NGBlockNode::FirstChild. This works
but grid has an additional placement cache which depends on if we
have a first child or not.

Invalidate the placement cache when the display lock context marks
a node for layout.

Bug: 1482765, 1482771
Change-Id: I3a47a192f938c4c14b3c2d1a868d78e30765c3db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4875138
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198308}
NOKEYCHECK=True
GitOrigin-RevId: 2846f3952d74a070d69bb4395af61b45ab413963