[@container] Fix ::highlight() style recalc for size containers
Highlight pseudo styles are updated along with the originating element's
style in RecalcOwnStyle(). If the originating element is a size
container and the highlight styles depend on size container queries, we
need to revisit the highlight style computation when the layout size is
known.

Bug: 1481068
Change-Id: Icc9b26565e2ffc5e0b06b51d9175ccb8648104de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4868848
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198359}
NOKEYCHECK=True
GitOrigin-RevId: 553bb8ea0a4b7aae4a6449aa1a8532bd04675528