[Track Stats] Implement the stats cache.
This CL makes it so that stats are only updated if this is the first
time in the task execution cycle that the getters are invoked,
preserving JS run-to-completion semantics, and otherwise returns
cached values.

This CL also moves the responsibility of maybe-updating into the
MediaStreamTrackVideoStats interface. This allows keeping direct
references to the stats object and calling videoStats.totalFrames etc
directly without first calling track.videoStats, as was previously
required (a bug).

WPTs for both features are added.

Bug: chromium:1472978
Change-Id: I8f2cddbb9b9d4a1a8ff5c3b11b22ea6613de2084
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4865101
Reviewed-by: Tony Herre <toprice@chromium.org>
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198917}
NOKEYCHECK=True
GitOrigin-RevId: 2ae669ed22c302b219a2e9f9d1e4bbd89ed5394a