Deflake MediaStreamTrack-video-stats.https.html.
A test was recently added which asserts that frames are frozen during a
task execution cycle. But it also asserted that after 100 ms and one
task execution cycle later, the frame counter would have incremented.

This is a reasonable assumption on a fast machine, but the WPT waterfall
can be really slow in which case no frame is produced in 100 ms and the
test flakily failed, see recent failures at
https://ci.chromium.org/ui/p/chromium/builders/ci/mac12-arm64-rel-tests

It's best to just remove this assertion. Other tests already cover
"wait for frame counters to increment" without a time limit, so there is
no need to relax it here and we don't want to increase the time that it
blocks...

NOTRY=True

Bug: chromium:1485075
Change-Id: If31cb74ce1ab69e795cdc4070601e3ea4afed3ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4876650
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198998}
NOKEYCHECK=True
GitOrigin-RevId: 0de6f0336e9004de563dab3ce429f02abf4197f4