[PeerConnection] Remove DCHECK from PeerConnectionTracker
This CL removes a DCHECK that asserted that a receiver
for a LocalDOMWindow is not yet bound when binding the receiver.
However, this can happen is some cases when a LocalDOMWindow is
reused for a different LocalFrame. See https://crbug.com/1484406#c1

Bug: 1484406
Change-Id: I3367988f73f22dbd42db317895e3e73e84326ee4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4877019
Auto-Submit: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Tony Herre <toprice@chromium.org>
Commit-Queue: Tony Herre <toprice@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199012}
NOKEYCHECK=True
GitOrigin-RevId: afdbb33ef0443a719e3ddb9c82a00582d51c75e8