Fix 'currentcolor' handling for drop-shadow(...) filter functions
For drop-shadow(), the 'currentcolor' keyword was resolved at style
resolution time by FilterOperationResolver. This means that it was baked
into style, and couldn't respond to a change in color.

Instead preserve the keyword in the FilterOperation and let the
FilterEffectBuilder resolve the color if needed. Add infrastructure to
FilterOperation/s to determine if a filter chain depends on
'currentcolor' to allow invalidation of the filter chain when the color
changes.

Bug: 1469327
Change-Id: I81bd0324637f1c1111add5d74c58418f205c3c77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4863914
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199067}
NOKEYCHECK=True
GitOrigin-RevId: a8c7ce4436274b29971e4a8a95522ab96c3ed088