[wptrunner] Run tests in a subprocess for `--use-upstream-wpt`
`--use-upstream-wpt` is unreliable because it relies on path hacking
to (re)load modules at runtime. This can be accomplished more robustly
by running tests in a fresh interpreter with PYTHONPATH pointing to the
`wpt/tools` to use. This eliminates the possibility that `blinkpy`
transitively imports a `wpt*` module that will be invalidated later.

When invoking the subprocess, the `--use-upstream-wpt` flag is stripped
away, but the `WPTAdapter` can still dynamically detect whether it's
running with Chromium-vended WPT.

Fixed: 1473716
Cq-Include-Trybots: luci.chromium.try:linux-wpt-content-shell-fyi-rel
Test: GTEST_TOTAL_SHARDS=100 run_wpt_tests.py --use-upstream-wpt
Change-Id: I5953403aa5233a8d29dec8f3c6c46dd0482d42a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4877821
Commit-Queue: Jonathan Lee <jonathanjlee@google.com>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Cr-Commit-Position: refs/heads/main@{#1199100}
NOKEYCHECK=True
GitOrigin-RevId: 5344cd65d075c179b750f28701985ac991099426