Set origin to commit for data: URLs
We need to set the origin_to_commit for data: URLs as part of the
transition to removing source SiteInstance from RenderFrameProxyHost.
Because data: URLs have opaque origins, we cannot identify which data:
URL is the source SiteInstance if there are multiple data: SiteInstances
in a group. By using the origin_to_commit, we can compare origin nonces
to make sure we use the correct data: URL SiteInstance as the source
SiteInstance.

We want the browser and renderer side nonces to match. Without this change, the
renderer calculates its own opaque origin, which will differ from the origin
used on the browser side to create a data: SiteInstance. If the browser and
renderer side origins don't match, we can't use it to identify a data: URL
SiteInstance.

We do not include LoadDataWithBaseUrl cases with a base URL because we don't
need SiteInstanceGroup in those cases. The base URL will be used to determine
the SiteInstance. If the base URL is empty, then the URL is treated like a
regular data: URL.

Bug: 1447896, 1291351
Test: Added NavigationControllerBrowserTest.DataURLHasCommittedOrigin
Change-Id: Ifaf3acfa15d16aff78b8643f88fcc69c619b5c5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4824705
Commit-Queue: Sharon Yang <yangsharon@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199187}
NOKEYCHECK=True
GitOrigin-RevId: 45a36540ff34c7ae4876a483f71e74e5a9e2b745