view-transition: Rename GetActiveTransition to GetTransition
The level 2 spec defines what it means to "activate" a view transition:
https://drafts.csswg.org/css-view-transitions-1/#activate-view-transition
This implies an "active" transition is one that's running on the
destination state of the page. (i.e. capturing elements or playing
the animation)

However GetActiveTransition returns an existing non-finished transition
which is a different meaning from the one in the spec.

This CL renames GetActiveTransition to GetTransition so that "active"
can be used for the meaning defined by the spec.

Bug: 1466250
Change-Id: I4ea60df5e60af298eed13ae5ac8fb8a2e3fb1499
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4873736
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199228}
NOKEYCHECK=True
GitOrigin-RevId: 830fcaef83e77a5b05963c1e73fb3d6ac6a4bc11