Simplify handling of overflowed fragmented block-in-inline.
Do it the same way as for a fragmented float in an inline formatting
context. An overflowed block-in-inline is very similar, since it too
establishes a parallel fragmentation flow. [1] Wrap the break token of
the block-in-inline inside an parallel block flow inline break token and
propagate it to the parent block layout algorithm via
NGInlineChildLayoutContext. Parallel flow break tokens inside an inline
formatting context will be picked up as ParallelFlowBreakTokens() in
NGBlockLayoutAlgorithm::FinishInflow().

[1] https://www.w3.org/TR/css-break-3/#parallel-flows

This simplifies the structure of NGInlineBreakToken, and requires less
code in NGLineBreaker and also elsewhere.

No behavior changes intended.

Change-Id: Ie07707670793453147a07c2c4e0d21fbfffebafa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4874029
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199426}
NOKEYCHECK=True
GitOrigin-RevId: 3287eab4c99198b96f7833b8d09e99956f9f1759