[task-attribution] Reland: Move to an implicit GCed task container model
The current task_container_ is a circular buffer with limited space.
That limits the potential use cases for TaskAttribution, and adds
needless complexity. This CL significantly simplifies that model, by
moving the container to be implicit, and relying on GC to clean up tasks
once they are no longer needed - that is, once all their descendent
tasks were done and evicted.

This is a reland of [1], that moves the MessagePort destructor work to
a prefinalizer, to avoid invalid memory reads during destruction.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/4614509

Change-Id: I2d61092f96cfd8207ca899c81c15ac808f094300
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4831380
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199538}
NOKEYCHECK=True
GitOrigin-RevId: b7d76ddd1b1d13ef27b31483bc64b51f0e1d5ce8