Blink: Remove old CanChangeToUrlForHistoryApi() code
This CL removes the feature guard code (and old, flag-guarded behavior)
for the history URL change code. It is a follow-up to
https://crrev.com/c/4701743.

Per
https://chromiumdash.appspot.com/commit/16b5623e6ef896e7e9b57546cd97881310e991ec,
this has baked in stable for a few weeks, since 117.0.5938.48, so we're
comfortable removing the flag guarding now and fully committing to this.

R=dcheng@chromium.org

Bug: 1239052
Change-Id: Ib44a92c856c2a18256372070b1dc6d7c60c8f485
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4833236
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199541}
NOKEYCHECK=True
GitOrigin-RevId: abc5f36163a1d9411135fc5591ba48119b89f7f5