Dictation: Query mic mute status before starting dictation
This change fixes the bug (if the feature flag [1] is enabled) by
querying the microphone mute status before starting Dictation. If
the microphone is muted, we don't start speech recognition and
instead show a toast notification in the bottom
right corner of the screen.

[1] enable-accessibility-dictation-keyboard-improvements

Bug: b:299677121
Change-Id: Ie703d6aa445e40bf681301eacda25bcfccd7d6e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4876022
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Andre Le <leandre@chromium.org>
Reviewed-by: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199937}
NOKEYCHECK=True
GitOrigin-RevId: d9e36d44fbf44cc7470897031af563f68845bc75