Support ignoring charset when parsing URLs in CSS behind a flag
This is to allow testing of the desired behavior, and it's a possible
cause of an issue being discussed in Sass:
https://github.com/sass/dart-sass/issues/568

There are two call sites for CSSParserContext::Charset(), one for the
regular parsing path, and one for parsing registered property values.
Rather than touching both, just throw away the charset override when the
feature is enabled.

Bug: 1485525

Change-Id: I3da69b4156d6a84bcc8e0517c954a79b522a9ec9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4846986
Commit-Queue: Philip Jägenstedt <foolip@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1200109}
NOKEYCHECK=True
GitOrigin-RevId: d9ae032e4c90fad3f18b9afb1d80755688fbf561