Send automatic beacons to all registered URLs.
Previously, automatic beacons would only send to destinations that were
explicitly specified when calling setReportEventDataForAutomaticBeacons.
That means that if the call did not specify a destination, but the
worklet for that destination registered an automatic beacon through
registerAdBeacon(), that destination would not get a beacon.

Any destination that asked to receive a beacon for a top-level
navigation has a vested interest in knowing if a something happened
based on a click. The frame that calls
setReportEventDataForAutomaticBeacons() should not be allowed to
interfere with whether a destination gets that information. However,
that frame should still have control over what destinations get the data
it’s setting for the automatic beacon.

This CL modifies that behavior so that if the above case happens, the
destination will now get sent an automatic beacon, but the automatic
beacon data will not be included in the payload. This will allow for
all top-level navigation events to notify every endpoint that wants to
be notified, but still give the frame control over who gets access to
the actual data.

As part of this change, the "once" parameter will now determine whether
the beacon data is sent once/multiple times, rather than the old
behavior of it determining if the beacon itself is sent once/multiple
times.

This change is gated behind the "FencedFramesM119Features" flag.

This CL also updates web platform tests and some fenced frame browser
tests to enable the "FencedFramesM119Features" flag, allowing for
testing this change.

See doc: https://docs.google.com/document/d/1vLifppH8TC86sl4kbam57egttqb4VNwEkVvZdefGIuQ/edit?usp=sharing

Change-Id: Iba4f4d56b2f5a2162cde2d3546c208f70dfd1abd
Bug: 1482328
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4860420
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Liam Brady <lbrady@google.com>
Reviewed-by: Garrett Tanzer <gtanzer@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1200208}
NOKEYCHECK=True
GitOrigin-RevId: 3d49f456c9b9af835217c401b2bcb25b70d4d51a