Cause the TimeZoneChange event to be enqueued
The implementation of the timezone change event was such that it
directly dispatched the event. This really should be an enqueue so that
this works well with paused pages. This matches the pending spec text
that the event is queued to the Window.

Bug: 908550
Change-Id: I76c0fb2492d1cd7ef1b95e98485f2b8968c80f02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4873567
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Frank Tang <ftang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1200224}
NOKEYCHECK=True
GitOrigin-RevId: 63050f161c9e11a30bb2157bbf29cd956b55361a