Speculative deflake of animation-inactive-outside-range
When flaking on a Mac, the ref image is wrong suggesting that the
screenshot is being taken too soon.
When flaking on Linux, an indicator light is incorrectly on in the
actual image.

The speculative fix, moves waitForComspoitorReady to the start of the
test and ref image so that all scrolling occurs after the first
contentful paint. The ref test now uses a wait to ensure completion of
the scroll (which happens on the compositor thread in some virtual
test suites.

Zero flakes in 500 runs on Linux. Do not have a Mac for testing.

Bug: 1481709
Change-Id: Ide770de6fe6aeba53339ae068e962159b47d2ea3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4873591
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1200264}
NOKEYCHECK=True
GitOrigin-RevId: aaa4183f8d672e541b6425adaa1f18198435d5ac