Implement rounded corner and native theme fluent changes.
When overlay scrollbars are in Full mode the buttons will have rounded
corners. This is in line with the visual specs and improves the
scrollbar's "floating" look given by the part's insets.

Web tests for fluent overlay scrollbars will be implemented in the
next CL.

Bug: 1479173
Change-Id: I3bc4ce592fd47fe2e027e582a8405f9601707093
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4876330
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1200542}
NOKEYCHECK=True
GitOrigin-RevId: 4e0aa6fc2bc912d69c2b0cacda2ff3ce1c1e5d16