[LCP] Make ImageRecord GarbageCollected
ImageRecord is currently not GarbageCollected, and in order to be able
to safely keep it around as part of [1], it should be.
On top of that, this CL removes a bunch of unneeded complexity around
|size_ordered_set_| and the |ImageRecordAndHashPair| struct. In part
because it was complex to implement (no heap-safe sorted set), but also
because we were paying O(log(n)) insertions for no good reason.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/4846835

Change-Id: Iff3b83d2278c4c740b199a5e40a50d3e63362e9d
Bug: 1479355
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4886912
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1200828}
NOKEYCHECK=True
GitOrigin-RevId: 868885b6574367e14695193d498a0ec2472c8342