Forward Capture process frame drops to Renderer process.
The MediaStreamTrack is currently unaware of what the camera's frame
rate is, because if a frame is dropped in the Capture process this is
never forwarded to the track. This is a problem for the implementation
of the MediaStreamTrack Statistics API, which currently counts delivered
frames correctly but not dropped frames.

The MediaStreamTrackImpl aims to count all frames that are dropped, but
is currently only aware of frames that were dropped in the Renderer
process. These frame drops are signalled to the Browser process'
VideoCaptureController where UMAs are counted.

The VideoCaptureController also received signals about frame drops in
the Capture process. This CL makes the controller forward frame drops
that originated in the Capture process to the Renderer process (but does
not forward frame drops that happened in the Renderer process, since the
Renderer process is already aware of these).

See diagram: https://bugs.chromium.org/p/chromium/issues/attachment?aid=613344&signed_aid=qgr6SdlcmdfuTpNMqHyzFQ==&inline=1

In the future, when MediaStreamTrackImpl is aware of all frame drops,
we can move the UMAs to MediaStreamTrackImpl and delete the frame drop
callbacks that are reported from Renderer -> Browser processes. This
would make all frame drop signals go in a single direction:
Capture -> Browser -> Renderer

In a follow-up CL we will continue to plumb the frame drop signal until
it reaches MediaStreamTrackImpl.

Bug: chromium:1472978
Change-Id: I623a4f90d3630ac1d736ee04b7d45be514887304
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4878697
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1200924}
NOKEYCHECK=True
GitOrigin-RevId: ce9c0569857596a5cbe8463b57c819088b16c704