SmartCard: split state flags into 'in' and 'out' versions
Instead of a SmartCardReaderStateFlags dictionary for both
SmartCardReaderStateIn and SmartCardReaderStateOut, have separate,
corresponding, SmartCardReaderStateFlagsIn and
SmartCardReaderStateFlagsOut dictionaries.

There's a lot of overlap between the two but they are *not* the same:
some values make sense only as input and some only as output.

It also makes it match the spec.

Bug: 1467264
Change-Id: I5f03f47e0be9073244bc9fcf200e96c79121d53b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4889638
Reviewed-by: Andrew Rayskiy <greengrape@google.com>
Commit-Queue: Daniel d'Andrada <dandrader@google.com>
Cr-Commit-Position: refs/heads/main@{#1200958}
NOKEYCHECK=True
GitOrigin-RevId: ae313e088174b26b68723aa4cc80746004f97836