Add frame dropped callback to StartCapture().
As of [1], if getUserMedia() tracks drop frames in the Capture process,
this is signalled to VideoCaptureImpl. In this CL, we plumb the frame
drop signal all the way to MediaStreamVideoSource by adding a
`frame_dropped_callback` parameter to StartCapture().

The relevant stack is updated to wire it down, while other types of
sources that don't have frame dropping can ignore the callback.

This is just an intermediate plumbing CL. The final wiring to the
MediaStreamTrackImpl happens in the next CL.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/4878697

Bug: chromium:1472978
Change-Id: I4b44a6e721154fefe3623a65e71c2cd0c99e1994
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4878700
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Reviewed-by: Fabian Sommer <fabiansommer@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1201063}
NOKEYCHECK=True
GitOrigin-RevId: 809dd111f5e273373da48df4f690843e0c4e5c4b