Wire up Capture process frame drops to MediaStreamVideoTrack.
This CL wires up the frame drop callback from the source (which is
aware of Capture process frame drops) to the VideoTrackAdapter, and
makes the VideoAdapter invoke DoNotifyFrameDropped() which is already
wired up to MediaStreamVideoTrack.

This means that the MediaStreamVideoTrack is now aware of all frame
drops, not just the frame drops happening inside VideoTrackAdapter. The
track treats all frame drops the same and does not care if it happened
early or late in the pipeline. This means it is included in the
MediaStreamTrack Statistics API.

This is unit test-covered using a fake source to signal frame drops.

As for end-to-end testing, I have also manually verified that calling
VideoFrameReceiver::OnFrameDropped() inside the Capture process
results in `track.stats.totalFrames` incrementing in JavaScript. This
is unfortunately not testable in Web Platform Tests because fake
sources are not available to JS, and under normal circumstances frames
are flowing without drops.

Bug: chromium:1472978
Change-Id: I472a6176704bd0e9d1c2540744372a242c0b37a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4878679
Commit-Queue: Henrik Boström <hbos@chromium.org>
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1201065}
NOKEYCHECK=True
GitOrigin-RevId: e225c835d9c9722532e829d779f7f2c9dd7b2119