May break before float, but still fit subsequent in-flow content.
It is possible for a float to be pushed to the next fragmentainer, while
still fitting subsequent in-flow content in the current fragmentainer.
Therefore, keep looking for content inside a culled inline, even if
we've found a float that belongs in a later fragmentainer.

This is similar to how a float may be pushed below the current line if
it doesn't fit beside it, while still laying out items that come after
it DOM-tree-wise, on the same line. A float may be pushed to the next
fragmentainer if it doesn't fit in the current one. But there may still
be room for plenty of in-flow lines in the current fragmentainer.

Bug: 1486411
Change-Id: Ic935dc9655d628ce33d2c5a0f20369ef5eba590f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4891619
Commit-Queue: Koji Ishii <kojii@chromium.org>
Auto-Submit: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1201198}
NOKEYCHECK=True
GitOrigin-RevId: c7cc12d52716ecb1f28f671eed42793a4a7fd501