IndexedDB: fix IndexedDBKey fuzzer stack overflow
There was a stack overflow when the input caused the fuzzer to
recurse excessively when creating very deeply nested (although not
infinite) IndexedDBKeys. The fuzzer is fixed to handle this case.

This also adds a check for `kMaximumDepth` to the code under
test, which is `content::EncodeIDBKey()`. This check had already
been enforced, but only in the renderer.

The browser already enforced the max depth in
`content::DecodeIDBKey()`.

A move constructor is also added to `IndexedDBKey`, without which
there was still an OOM due to excessive copying, even after fixing
the stack overflow.

Bug: 1485116
Change-Id: I8ade9d02d513f2478e4b0a18227921e7c1e97827
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4892130
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Austin Sullivan <asully@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1201224}
NOKEYCHECK=True
GitOrigin-RevId: 2f68908b635bddd157828bf66d22296c063dc89d